angular ui skeleton added, migrating to postgres, flyway added, some fixes

+++ src/main/java/eins/dao/CompanyUserDao.java
@@ -4,6 +4,6 @@ import eins.entity.CompanyUser;
 import eins.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyUserDao extends JpaRepository<CompanyUser,Integer> {
+public interface CompanyUserDao extends JpaRepository<CompanyUser, Long> {
 
 }
