+++ src/main/java/eins/dao/ProductToBuyDao.java
@@ -3,7 +3,7 @@ package eins.dao;
 import eins.entity.ProductToBuy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductToBuyDao extends JpaRepository<ProductToBuy,Integer> {
+public interface ProductToBuyDao extends JpaRepository<ProductToBuy, Long> {
 
 
 }
